relocate sample unit and instrumentation tests to corresponding modules
