+++ di/src/androidTest/java/com/mustafakocer/di/ExampleInstrumentedTest.kt
@@ -1,4 +1,4 @@
-package com.mustafakocer.database_contracts
+package com.mustafakocer.di
 
 import androidx.test.platform.app.InstrumentationRegistry
 import androidx.test.ext.junit.runners.AndroidJUnit4
@@ -19,6 +19,6 @@ class ExampleInstrumentedTest {
     fun useAppContext() {
         // Context of the app under test.
         val appContext = InstrumentationRegistry.getInstrumentation().targetContext
-        assertEquals("com.mustafakocer.database_contracts.test", appContext.packageName)
+        assertEquals("com.mustafakocer.di.test", appContext.packageName)
     }
 }
